chainlit/requirements.txt
python-dotenv

pyodbc
azure-identity

langchain 
langchain-openai
langchain-community 
langchain-sqlserver
langgraph

chainlit


